[ actions ] Relaxing build constraints (#8)
We are now building the project with all of the versions of ghc made available
to us by https://github.com/actions/setup-haskell

To make the builds as fast as possible, we are caching `.cabal` and `.ghc`
for each one of the versions. So far that should take around 1G (5*200M)
which is way below the 5G allowance we have.