[ new ] Import located based on the current file
Rather than based on the current working directory when invoking
mary...

If the import is not found in the current directory of the file, we
move one step closer to the root and try again. Hence if
/a/b/c/file.shonkier imports "d/import.shonkier", we look for in order
/a/b/c/d/import.shonkier
/a/b/d/import.shonkier
/a/d/import.shonkier
/d/import.shonkier

Global variables and environments now contain absolute paths, but the
greatest longer prefix of each paths is stripped off.